requirements.txt
icrawler
opencv-python
mediapipe
torch
torchvision
numpy
ultralytics
seaborn
pandas
tqdm
pyyaml
gitpython
matplotlib
scipy
Pillow
requests
streamlit
streamlit-webrtc
av
plotly